ai-backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
fastapi==0.115.12
flatbuffers==25.2.10
gast==0.6.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.9.0
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
namex==0.0.9
numpy==2.1.3
openai==1.76.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==10.4.0
proglog==0.1.11
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
termcolor==3.0.1
tf_keras==2.19.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
